refactor(sdk): extract shared Erlang SDK construction helper and move SdkProcessUtils to util
 SdkRegistrar.registerOrUpdateErlangSdk and the synchronous EDT path in ElixirInternalErlangSdkSetup.registerErlangSdk both constructed a ProjectJdkImpl from the same inputs: canonicalised home path, version string, and suggested SDK name. Any drift between the two paths would silently produce SDKs with different names for the same Erlang installation. The EDT path also had no dedup - it always called addJdk even if the SDK was already registered at that home path.

 Extract ErlangSdkPreparation (existing: Sdk?, template: ProjectJdkImpl) and prepareErlangSdk() into SdkRegistrar as an internal helper. registerOrUpdateErlangSdk delegates to it. registerErlangSdk now also uses it, gaining home-path dedup and eliminating the duplicated construction logic.

 SdkProcessUtils lives in org.elixir_lang.sdk but contains no SDK logic - it is a general threading utility. Move it to org.elixir_lang.util as ProgressUtils.kt alongside the other threading helpers.
