+++ src/org/elixir_lang/sdk/SdkRegistrar.kt
@@ -12,21 +12,35 @@ import org.elixir_lang.sdk.elixir.Type as ElixirSdkType
 import org.elixir_lang.sdk.erlang.Type as ErlangSdkType
 import org.elixir_lang.sdk.erlang_dependent.Type as ErlangDependentType
 
+/** Holds the pre-computed parts for registering or updating an Erlang SDK. */
+internal data class ErlangSdkPreparation(
+    val existing: Sdk?,
+    val template: ProjectJdkImpl,
+)
+
 object SdkRegistrar {
+    /**
+     * Canonicalises [homePath], resolves the version string and SDK name, and looks up any
+     * already-registered SDK at that home. Returns null if the path or version cannot be resolved.
+     * Shared by both the suspend ([registerOrUpdateErlangSdk]) and synchronous
+     * ([org.elixir_lang.sdk.elixir.ElixirInternalErlangSdkSetup.registerErlangSdk]) registration paths.
+     */
+    internal fun prepareErlangSdk(homePath: String, resolvedVersion: String? = null): ErlangSdkPreparation? {
+        val canonicalHomePath = canonicalHomePath(homePath) ?: return null
+        val sdkType = ErlangSdkType.instance
+        val versionString = ErlangSdkType.versionStringForHome(canonicalHomePath, resolvedVersion) ?: return null
+        val sdkName = ErlangSdkType.suggestSdkNameForHome(canonicalHomePath, resolvedVersion)
+        val existing = findSdkByHomePath(sdkType, canonicalHomePath)
+        return ErlangSdkPreparation(existing, ProjectJdkImpl(sdkName, sdkType, canonicalHomePath, versionString))
+    }
+
     suspend fun registerOrUpdateErlangSdk(
         homePath: String,
         resolvedVersion: String? = null,
     ): Sdk? {
-        val canonicalHomePath = canonicalHomePath(homePath) ?: return null
-        val sdkType = ErlangSdkType.instance
-        val versionString = ErlangSdkType.versionStringForHome(canonicalHomePath, resolvedVersion) ?: return null
-        val existing = findSdkByHomePath(sdkType, canonicalHomePath)
-        val sdkName = ErlangSdkType.suggestSdkNameForHome(canonicalHomePath, resolvedVersion)
-
-        val updatedSdk = ProjectJdkImpl(sdkName, sdkType, canonicalHomePath, versionString)
-        val registeredSdk = registerOrUpdate(existing, updatedSdk)
-
-        sdkType.setupSdkPaths(registeredSdk)
+        val prep = prepareErlangSdk(homePath, resolvedVersion) ?: return null
+        val registeredSdk = registerOrUpdate(prep.existing, prep.template)
+        ErlangSdkType.instance.setupSdkPaths(registeredSdk)
         return registeredSdk
     }
